Refactor integration test to use copy_fixture pattern
Replace inline fixture construction with a static fixture directory
under tests/fixtures/custom-rule-config/ and copy_fixture(name, tmp_path),
matching the test suite convention.
