+++ tests/fixtures/custom-rule-config/.skillsaw.yaml
@@ -1,0 +1,2 @@
+custom-rules:
+  - .skillsaw-custom.py
