+++ tests/fixtures/custom-rule-config/plugins/my-plugin/.claude-plugin/plugin.json
@@ -1,0 +1,1 @@
+{"name": "my-plugin", "description": "d", "version": "1.0.0", "author": {"name": "a"}}
