+++ tests/fixtures/custom-rule-config/plugins/my-plugin/commands/hello.md
@@ -1,0 +1,5 @@
+---
+description: hello
+---
+
+# hello
